+++ src/CVMap.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include "plugin.hpp"
+
+namespace StoermelderPackOne {
+namespace CVMap {
+
+struct CVMapCtxBase : Module {
+	virtual std::string getCVMapId() { return ""; }
+};
+
+} // namespace CVMap
+} // namespace StoermelderPackOne